- feat: add originalIndex property to use in link generatior - feat: add placeholder error banner for malformed external url - feat: when users upload cohort, switch to the first samples - refactor: reorganize components/files
