+++ src/ui/UploadModal/index.tsx
@@ -1,9 +1,10 @@
 import React, { useState, useMemo } from 'react';
-import { ICONS } from '../icon';
-import { FileDragUpload } from './VisOverviewPanel/FileDragUpload';
-import { SampleConfig, ValidCohort } from './SampleConfigForm';
-import { Cohorts, Cohort } from '../App';
-import { ValidSampleConfig } from './SampleConfigForm';
+import { ICONS } from '../../icon';
+import { FileDragUpload } from '../VisOverviewPanel/FileDragUpload';
+import { SampleConfig, ValidCohort } from '../SampleConfigForm';
+import { Cohorts, Cohort } from '../../App';
+import { ValidSampleConfig } from '../SampleConfigForm';
+import { UploadModalFeedback } from './UploadModalFeedback';
 
 const exampleConfigFields: SampleConfig = {
     id: '7a921087-8e62-4a93-a757-fd8cdbe1eb8f',
@@ -13,169 +14,13 @@ const exampleConfigFields: SampleConfig = {
     cnv: 'https://s3.amazonaws.com/gosling-lang.org/data/SV/7a921087-8e62-4a93-a757-fd8cdbe1eb8f.consensus.20170119.somatic.cna.annotated.txt'
 };
 
-type UploadModalFeedbackProps = {
-    uploadedFile: File | null;
-    uploadedFileData: any;
-    uploadedCohort: ValidCohort;
-    setUploadedCohort: React.Dispatch<React.SetStateAction<ValidCohort>>;
-    error: string | null;
-    setError: React.Dispatch<React.SetStateAction<string | null>>;
-};
-
-const UploadModalFeedback = ({
-    uploadedFile,
-    uploadedFileData,
-    uploadedCohort,
-    setUploadedCohort,
-    error,
-    setError
-}: UploadModalFeedbackProps) => {
-    const [showAllSamples, setShowAllSamples] = useState(false);
-    const [cohortName, setCohortName] = useState(uploadedCohort?.name);
-    const [changingCohortName, setChangingCohortName] = useState(false);
-
-    // Handle cohort name change
-    const handleCohortNameChange = (newName: string) => {
-        if (newName && newName.length > 0) {
-            setUploadedCohort &&
-                setUploadedCohort({
-                    ...uploadedCohort,
-                    name: newName
-                });
-            setChangingCohortName(false);
-        }
-    };
-
-    if (error) {
-        return (
-            <div className="upload-feedback">
-                <div className="feedback-banner error">
-                    <div className="header">
-                        {/* <svg className="" viewBox={ICONS.ERROR.viewBox}>
-                            {ICONS.ERROR.path.map(p => (
-                                <path fill="currentColor" key={p} d={p} />
-                            ))}
-                        </svg> */}
-                        <span>Upload Failed</span>
-                    </div>
-                    <div className="body">
-                        <span>{error}</span>
-                    </div>
-                </div>
-            </div>
-        );
-    }
-
-    return (
-        <div className="upload-feedback">
-            <div className="feedback-banner success">
-                <div className="header">
-                    <svg className="" viewBox={ICONS.UPLOAD_FILE.viewBox}>
-                        {ICONS.UPLOAD_FILE.path.map(p => (
-                            <path fill="currentColor" key={p} d={p} />
-                        ))}
-                    </svg>
-                    <span>Upload Successful: {uploadedFile?.name}</span>
-                </div>
-                <div className="body">
-                    <div className="summary">
-                        <div className="name">
-                            <b>Cohort Name:</b>
-                            {changingCohortName ? (
-                                <div className="name-input">
-                                    <input
-                                        id="cohort-name-input"
-                                        type="text"
-                                        // placeholder={uploadedCohort?.name || uploadedFile?.name}
-                                        value={cohortName || uploadedCohort?.name}
-                                        onChange={e => setCohortName(e.target.value || uploadedCohort?.name)}
-                                        onKeyPress={e => {
-                                            if (e.key === 'Enter') {
-                                                handleCohortNameChange(cohortName);
-                                                setCohortName(cohortName);
-                                            } else if (e.key === 'Esc') {
-                                                e.preventDefault();
-                                                setCohortName(uploadedCohort?.name);
-                                                setChangingCohortName(false);
-                                            }
-                                        }}
-                                    />
-                                    <button
-                                        className="cancel"
-                                        onClick={() => {
-                                            console.log('setting cohortName', cohortName, uploadedCohort?.name);
-                                            setCohortName(uploadedCohort?.name);
-                                            setChangingCohortName(false);
-                                        }}
-                                    >
-                                        <svg viewBox={ICONS.CLOSE.viewBox}>
-                                            {ICONS.CLOSE.path.map(p => (
-                                                <path fill="currentColor" key={p} d={p} />
-                                            ))}
-                                        </svg>
-                                    </button>
-                                    <button
-                                        className="save"
-                                        type="submit"
-                                        onClick={() => {
-                                            handleCohortNameChange(cohortName);
-                                        }}
-                                    >
-                                        Save
-                                    </button>
-                                </div>
-                            ) : (
-                                <div className="name-display">
-                                    <span>{uploadedCohort?.name || uploadedFile?.name}</span>
-                                    <button
-                                        className="edit"
-                                        onClick={() => {
-                                            setCohortName(cohortName);
-                                            setChangingCohortName(true);
-                                        }}
-                                    >
-                                        Edit
-                                        {/* <svg viewBox={ICONS.PENCIL.viewBox}>
-                                            {ICONS.PENCIL.path.map(p => (
-                                                <path fill="currentColor" key={p} d={p} />
-                                            ))}
-                                        </svg> */}
-                                    </button>
-                                </div>
-                            )}
-                        </div>
-                        <span className="">
-                            <b>{uploadedFileData?.samples ? uploadedFileData.samples.length : 0} Samples Uploaded</b> -{' '}
-                            <button onClick={() => setShowAllSamples(!showAllSamples)}>
-                                <i>{showAllSamples ? 'Hide' : 'View All'}</i>
-                            </button>
-                            <br />
-                        </span>
-                    </div>
-                    <div className="samples">
-                        {showAllSamples && (
-                            <ul>
-                                {uploadedFileData?.samples.map((sample: any, key: number) => (
-                                    <li key={sample.id}>
-                                        <span>{sample.id}</span>
-                                    </li>
-                                ))}
-                            </ul>
-                        )}
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-};
-
 /**
  * Checks if a string is a valid URL.
  * @param str String to test as URL
  * @returns `true` if the string is a valid URL, `false` otherwise.
  */
 export function isValidUrl(str: string) {
-    let url;
+    let url: string | URL;
     try {
         url = new URL(str);
     } catch (_) {
@@ -206,6 +51,8 @@ type UploadModalProps = {
     cohorts: Cohorts;
     uploadedCohort: ValidCohort;
     selectedCohort: string;
+    demoIndex: React.MutableRefObject<number>;
+    setDemo: (demo: SampleConfig) => void;
     setSelectedCohort: (cohort: string) => void;
     setUploadedCohort: React.Dispatch<React.SetStateAction<ValidCohort>>;
     setCohorts: (cohorts: Cohorts) => void;
@@ -225,6 +72,8 @@ export const UploadModal = ({
     setUploadedCohort,
     cohorts,
     selectedCohort,
+    demoIndex,
+    setDemo,
     setSelectedCohort,
     setCohorts,
     setSampleConfig,
@@ -318,28 +167,20 @@ export const UploadModal = ({
      * @param samples {ValidSampleConfig[]}
      */
     const createNewCohortWithSamples = (cohortId: string, samples: ValidSampleConfig[]) => {
-        if (!cohorts?.[cohortId]) {
-            // Create new cohort
-            setCohorts({
-                ...cohorts,
-                [cohortId]: {
-                    name: uploadedCohort?.name || cohortId,
-                    samples: samples
-                }
-            });
-            setSelectedCohort(cohortId);
-        } else {
-            const newCohortId = cohortId + '_1';
-            setCohorts({
-                ...cohorts,
-                [newCohortId]: {
-                    name: newCohortId,
-                    samples: samples
-                }
-            });
-            setSelectedCohort(newCohortId);
-        }
+        // Create new cohort ID if the given one already exists
+        const newCohortId = !cohorts?.[cohortId] ? cohortId : cohortId + '_1';
+
+        setCohorts({
+            ...cohorts,
+            [newCohortId]: {
+                name: uploadedCohort?.name || newCohortId,
+                samples: samples
+            }
+        });
+        setSelectedCohort(newCohortId);
         setUploadedCohort(null);
+        setDemo(samples[0]); // Set first sample as demo
+        demoIndex.current = 0;
     };
 
     // Check if a list of samples are valid
